backend/requirements_fastapi.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
